backend/pyproject.toml
[project]
name = "task-flow-api"
version = "0.1.0"
description = ""
authors = [
    {name = "David Schnoll",email = "davidschnoll@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "environs[django] (>=14.5.0,<15.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.9,<4.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg (>=3.2.12,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "httpie (>=3.2.4,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "task_flow_api.settings"
# -- recommended but optional:
python_files = ["tests.py", "test_*.py", "*_test.py", "testing/python/*.py"]
